Rename indicies.js to indexes.js
